requirements.txt
asgiref==3.8.1
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-phonenumber-field==8.0.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
packaging==24.1
phonenumbers==8.13.43
pillow==10.4.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
